refactor: 旧プロフィールサービスを削除しAPIルートを移行
