+++ app/routes/api/forms/create/route.tsx
@@ -8,8 +8,8 @@ import {
   ERROR_STATUS_MAP,
 } from "~/constants/errors";
 import { FormInputSchema } from "~/models/forms";
+import { getProfileByUserId } from "~/services/profiles/get.server";
 import { createServerSupabaseClient } from "~/services/supabase/client.server";
-import { getProfileByUserId } from "~/services/supabase/profiles";
 import { hasModeratorPermission } from "~/utils/permissions";
 
 import type { Route } from "./+types/route";
@@ -54,7 +54,7 @@ export const action = async (args: Route.ActionArgs) => {
   const { title, description, status } = submission.value;
 
   const supabase = createServerSupabaseClient(args);
-  const profileResponse = await getProfileByUserId(supabase, userId);
+  const profileResponse = await getProfileByUserId(args, userId);
 
   if (profileResponse.error || !profileResponse.data) {
     return data(
@@ -62,7 +62,9 @@ export const action = async (args: Route.ActionArgs) => {
         success: false,
         error: {
           code: ERROR_CODES.PROFILE_NOT_FOUND,
-          message: profileResponse.error || ERROR_MESSAGES_MAP[ERROR_CODES.PROFILE_NOT_FOUND],
+          message:
+            profileResponse.error ||
+            ERROR_MESSAGES_MAP[ERROR_CODES.PROFILE_NOT_FOUND],
         },
       },
       { status: ERROR_STATUS_MAP[ERROR_CODES.PROFILE_NOT_FOUND] },
